Tooltips to show variable details like state/type/signature/value.
The plug-in now shows tooltips for the current split window in graphical
Vim, when you hover over variables with the mouse cursor. When you don't
understand/can't explain the dynamic highlighting being performed by
luainspect.vim then just hover over the variable in question and you'll
get a textual description of whatever is causing the problem. When the
value of the variable is known you'll get a preview of that as well (in
the case of table the first 20 keys are shown and for functions the file
and line number where they were defined are shown).