Bug fix: Show error highlighting only in Lua buffers
When luainspect.vim requests a tool tip text from luainspect4vim.lua but
a syntax error is found then no tool tips are available, however the
syntax error is reported so that it can be highlighted as such. This
feature and its sequence of events introduced the following bug:

 1. Open a Lua buffer.
 2. Create a syntax error in the Lua buffer.
 2. Create a split window showing a non-Lua buffer.
 3. With the focus on the non-Lua buffer, hover over the split window
    showing the Lua buffer and wait for the tool tip to appear.
 4. The tool tip will be correct but the syntax error highlighting is
    now applied in the non-Lua buffer: Clearly not the intention :-)

This commit fixes the problem. I've also added calls to my xolox#timer
functions so that, when you execute the following command:

    :verbose LuaInspect

You get a message with the time spent inside luainspect4vim.lua and the
overhead imposed by the communication between Vim and Lua.