Ignore LuaInspect status messages on stderr
LuaInspect recently started emitting `status:' and `warning:' messages
to `stderr' which causes problems when running `luainspect4vim.lua' as
an external process because the Vim plug-in sees those messages as
output from `luainspect4vim.lua' and tries to parse them.

Ideally the Vim plug-in should just ignore all output on `stderr' but
that requires changing the API presented by my `shell.vim' plug-in and
this means I also have to update every other plug-in that integrates
with `shell.vim'... For this reason `luainspect.vim' now explicitly
ignores lines starting with `status:' or `warning:'.