Bug fix: Focus doesn't return to original window
When there's already a split window above the Lua buffer (e.g. a help
buffer) and the plug-in opens or updates the location list for the Lua
buffer it switches to the wrong buffer afterwards, which causes Vim to
emit a nasty stream of errors, each of which causes an individual
hit-enter prompt :-( -- very annoying.